oracle/agent/requirements.txt
langgraph>=0.2
langchain-anthropic>=0.3
langchain-core>=0.3
copilotkit>=0.1
ag-ui-langgraph>=0.0.20
redis[hiredis]>=5.0
pydantic>=2.0
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0
